Improved function signatures and refactored
Instead of passing storage_path and filename to functions. The note
struct, which was previously called struct write_info, contains all the
required members. Now, only a pointer to struct note is passed to functions.

The controller function is also refactored, and instead of freeing
memory from multiple switch cases manually, the cleanup function is now
responsible to handle the freeing of memory.