Added an option that disables writing the date
Added the "-d" option which set the write_date flag to false. So, no
date is written to the note file. Also, changed the argument parsing in
the parse_args function. Options except "-d" set the request_set flag
to true, so that the following options don't change the previous request
type.