Merged config_parse into config
Moved all the function from config_parser to config.c. The parse_cleanup
function is now merged with the cleanup function is notes.c. The idents
struct is moved to the header file config.h. Editor and date_fmt idents
will be enabled in the next commit. Project level refactoring is also
needed.