Refactored the project
The note struct now contains pointers to struct request and config. All
the note struct member uses are updated. The EDITOR environment variable
was not being used in the config, which is now fixed.

The config file is always being read and parsed which is reducing the
speed of the program. On succesful parsing the value will be stored in a
local database and the config file will only be read when modified. This
update will continue in the "database" branch. All the refactoring and
commenting will be done in the "dev" branch. "dev" branch will be merged
into "master" after all the tests are successful.